Adds release publication check to build jobs
Ensures that the build, pack, login, and push jobs are only executed when a new release is published.

This prevents unnecessary build processes when only a dry-run release is performed.

Also adds conventional-changelog-conventionalcommits and Release Notes Generator to semantic release.
Additionally outputs the state of "new_release_published" for debugging and future use cases.